Implemented Neumaier sum using PreciseSums package.
Seems to be just as accurate/stable as fsum, slightly faster.
This implementation leaves a lot ot be desired though. CPU
efficiency is pretty poor as the matrix addition addition is being
done with a nested for loop accross the dimensions of beta and sigma.
This should be done in parallel. In addition, I am summing all columns
accross all documents.